Fixed lq links containing 1080p episodes instead of 480p
Now that I think about it, it should have been HQ instead of LQ since HQ is how 480p its usually refer to. 
At this point, I don't feel like changing it tho, since I have to change a lot of files or it would broke all the app and the database (or both)... 

So lets just say that I never thought and/or had no idea about it.